refactor: 从annotation抽离card逻辑，封装到Card.ts
